+++ geotrek/maintenance/filters.py
@@ -102,6 +102,10 @@ class ProjectIntersectionFilterDistrict(PolygonInterventionFilterMixin, RightFil
 class InterventionFilterSet(ZoningFilterSet, StructureRelatedFilterSet):
     ON_CHOICES = (('infrastructure', _("Infrastructure")), ('signage', _("Signage")), ('blade', _("Blade")),
                   ('topology', _("Path")), ('trek', _("Trek")), ('poi', _("POI")), ('service', _("Service")),
+                  ('trail', _("Trail")), ('course', _("Course")), ('site', _("Site"))) \
+        if 'geotrek.outdoor' in settings.INSTALLED_APPS else \
+                 (('infrastructure', _("Infrastructure")), ('signage', _("Signage")), ('blade', _("Blade")),
+                  ('topology', _("Path")), ('trek', _("Trek")), ('poi', _("POI")), ('service', _("Service")),
                   ('trail', _("Trail")))
     bbox = PolygonTopologyFilter(lookup_expr='intersects')
     year = MultipleChoiceFilter(choices=Intervention.objects.year_choices(),