+++ geotrek/maintenance/tests/test_filters.py
@@ -17,6 +17,9 @@ from geotrek.maintenance.filters import ProjectFilterSet, InterventionFilterSet
 from geotrek.maintenance.factories import InterventionFactory, ProjectFactory
 from geotrek.zoning.factories import CityFactory, DistrictFactory
 
+if 'geotrek.outdoor' in settings.INSTALLED_APPS:
+    from geotrek.outdoor.factories import SiteFactory, CourseFactory
+
 
 class InterventionFilteringByBboxTest(TestCase):
     @classmethod
@@ -51,6 +54,13 @@ class InterventionFilteringByLandTest(TestCase):
 
         InterventionFactory.create(target=topo_1)
         cls.seek_inter = InterventionFactory.create(target=seek_topo)
+
+        if 'geotrek.outdoor' in settings.INSTALLED_APPS:
+            site = SiteFactory()
+            cls.seek_site = InterventionFactory.create(target=site)
+
+            outdoor = CourseFactory()
+            cls.seek_course = InterventionFactory.create(target=outdoor)
 
     def test_filter_by_physical_edge(self):
         edge = PhysicalEdgeFactory(paths=[self.seek_path])
@@ -105,6 +115,22 @@ class InterventionFilteringByLandTest(TestCase):
 
         self.assertEqual(len(qs), 1)
         self.assertEqual(qs[0], self.seek_inter)
+
+    @skipIf('geotrek.outdoor' not in settings.INSTALLED_APPS, 'Outdoor module not installed')
+    def test_filter_by_target_site(self):
+        # filter by target
+        data = {'on': 'site'}
+        qs = InterventionFilterSet(data=data).qs
+        self.assertEqual(len(qs), 1)
+        self.assertEqual(qs[0], self.seek_site)
+
+    @skipIf('geotrek.outdoor' not in settings.INSTALLED_APPS, 'Outdoor module not installed')
+    def test_filter_by_target_course(self):
+        # filter by target
+        data = {'on': 'course'}
+        qs = InterventionFilterSet(data=data).qs
+        self.assertEqual(len(qs), 1)
+        self.assertEqual(qs[0], self.seek_course)
 
 
 class ProjectFilteringByYearTest(TestCase):